Moves sentinel_{request,response}.rs into user_message.rs
And first take on simplifying handling of method calls
