+++ src/sentinel_response.rs
@@ -15,68 +15,4 @@
 // Please review the Licences for the specific language governing permissions and limitations
 // relating to use of the SAFE Network Software.
 
-use authority::Authority;
-use messages::{RoutingMessage, ErrorReturn, GetDataResponse};
-use name_type::NameType;
-use sentinel::pure_sentinel::Source;
-use types::MessageId;
 
-#[derive(Clone, Ord, PartialOrd, Eq, PartialEq)]
-pub struct SentinelPutResponse {
-    pub response: ErrorReturn,
-    pub source_group: NameType,
-    pub destination_group: NameType,
-    pub source_authority: Authority,
-    pub our_authority: Authority,
-    pub message_id: MessageId
-}
-
-impl SentinelPutResponse {
-    pub fn new(message: RoutingMessage, response: ErrorReturn, our_authority: Authority)
-        -> SentinelPutResponse {
-        SentinelPutResponse {
-            response: response,
-            source_group: message.source.non_relayed_source(),
-            destination_group: message.destination.non_relayed_destination(),
-            source_authority: message.authority,
-            our_authority: our_authority,
-            message_id: message.message_id
-        }
-    }
-}
-
-impl Source<NameType> for SentinelPutResponse {
-    fn get_source(&self) -> NameType {
-        self.source_group.clone()
-    }
-}
-
-#[derive(Clone, Ord, PartialOrd, Eq, PartialEq)]
-pub struct SentinelGetDataResponse {
-    pub response: GetDataResponse,
-    pub source_group: NameType,
-    pub destination_group: NameType,
-    pub source_authority: Authority,
-    pub our_authority: Authority,
-    pub message_id: MessageId
-}
-
-impl SentinelGetDataResponse {
-    pub fn new(message: RoutingMessage, response: GetDataResponse, our_authority: Authority)
-        -> SentinelGetDataResponse {
-        SentinelGetDataResponse {
-            response: response,
-            source_group: message.source.non_relayed_source(),
-            destination_group: message.destination.non_relayed_destination(),
-            source_authority: message.authority,
-            our_authority: our_authority,
-            message_id: message.message_id
-        }
-    }
-}
-
-impl Source<NameType> for SentinelGetDataResponse {
-    fn get_source(&self) -> NameType {
-        self.source_group.clone()
-    }
-}
